+++ package.json
@@ -10,6 +10,7 @@
   },
   "dependencies": {
     "next": "14.2.4",
+    "next-auth": "^5.0.0-beta.19",
     "react": "^18",
     "react-dom": "^18",
     "return-fetch": "^0.4.6"
