+++ src/app/api/auth/[...nextauth]/route.js
@@ -1,0 +1,3 @@
+import { handlers } from "@/auth";
+
+export const { GET, POST } = handlers;
