+++ src/app/login/page.js
@@ -1,76 +1,10 @@
-"use client";
-
-import { mFetch } from "@/util/MFetch";
+import { LoginForm } from "@/components/login";
 
 const LoginPage = () => {
-  const login = async (e) => {
-    e.preventDefault();
-    // mFetch("/user", {
-    //   credentials: "include",
-    //   headers: {
-    //     Authorization:
-    //       "Basic " + btoa(`${e.target.email.value}:${e.target.password.value}`),
-    //   },
-    // })
-    fetch("/api/login", {
-      method: "POST",
-      body: JSON.stringify({
-        email: e.target.email.value,
-        password: e.target.password.value,
-      }),
-    })
-      .then((response) => {
-        if (response.status === 200) {
-          return response.json();
-        } else {
-          throw new Error("Invalid credentials");
-        }
-      })
-      .then((data) => {
-        window.location.href = "/";
-      })
-      .catch((error) => {
-        console.error("Error:", error);
-      });
-  };
-
   return (
     <div className="bg-gray-100">
       <div className="flex justify-center h-screen w-screen items-center">
-        <form
-          onSubmit={login}
-          className="w-full md:w-1/3 flex flex-col items-center"
-        >
-          <h1 className="text-center text-2xl font-bold text-gray-600 mb-6">
-            LOGIN
-          </h1>
-          <div className="w-3/4 mb-6">
-            <input
-              type="email"
-              name="email"
-              id="email"
-              className="w-full py-4 px-8 bg-slate-200 placeholder:font-semibold rounded hover:ring-1 outline-blue-500"
-              placeholder="User Name or Email ID"
-            />
-          </div>
-          <div className="w-3/4 mb-6">
-            <input
-              type="password"
-              name="password"
-              id="password"
-              className="w-full py-4 px-8 bg-slate-200 placeholder:font-semibold rounded hover:ring-1 outline-blue-500 "
-              placeholder="Password"
-            />
-          </div>
-          <div className="w-3/4 mt-4">
-            <button
-              type="submit"
-              className="py-4 bg-blue-400 w-full rounded text-blue-50 font-bold hover:bg-blue-700"
-            >
-              Login
-            </button>
-          </div>
-        </form>
+        <LoginForm />
       </div>
     </div>
   );
