+++ src/util/MFetch.js
@@ -6,14 +6,10 @@ export const mFetch = returnFetch({
   baseUrl: apiUrl,
   interceptors: {
     request: async (args) => {
-      console.log("********* before sending request *********");
-      console.log(args);
       return args;
     },
 
     response: async (response, requestArgs) => {
-      console.log("********* after receiving response *********");
-      console.log(response);
       return response;
     },
   },
